api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
alembic==1.14.0
python-dotenv==1.2.1
websockets==13.1
kafka-python==2.3.0
python-snappy==0.7.3
